// afi-manager-base-model/model/HMTracking.ts
// Generated by https://quicktype.io

export class HMTracking {
  header?: Header;
  body?: Body[];
}

export interface Body {
  location?: string;
  timestamp?: string;
  status?: string;
  status_text?: string;
  status_details?: string;
  shown?: boolean;
  full_courier_status?: string;
  Specifics: string;
  created?: string;
}

export interface Header {
  id: string;
  tracking_number: string;
  actionBox: ActionBox;
  courier: Courier;
  last_delivery_status: LastDeliveryStatus;
  delay: boolean;
  exception: boolean;
  delivery_info: DeliveryInfo;
}

export interface ActionBox {
  type: string;
  label: string;
}

export interface Courier {
  name: string;
  prettyname: string;
  trackingurl: string;
  hide_trackingurl: boolean;
  trackingurl_label: string;
  rerouteurl: string;
  rerouteurl_label_short: string;
  rerouteurl_label_long: string;
  rerouteurl_label_info: string;
  destination_courier: DestinationCourier;
}

export interface DestinationCourier {
  name: string;
  prettyname: string;
}

export interface DeliveryInfo {
  isReturn: boolean;
}

export interface LastDeliveryStatus {
  code: string;
  status: string;
  status_details: string;
}